requirements.txt
fitz==0.0.1.dev2
Flask==3.1.0
huggingface_hub==0.27.0
python-dotenv==1.0.1
transformers==4.47.1
Werkzeug==3.1.3